Move controller resolver to controller namespace.
