+++ src/Synapse/Controller/ServiceProvider.php
@@ -15,7 +15,7 @@ class ServiceProvider implements ServiceProviderInterface
     public function register(Application $app)
     {
         $app['resolver'] = $app->share($app->extend('resolver', function ($resolver, $app) {
-            return new ControllerResolver($resolver, $app);
+            return new Resolver($resolver, $app);
         }));
 
         $app->initializer(
